pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-multi-worker"
version = "0.1.0"
description = "Document-Driven Multi-Agent Development Orchestration Framework"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["claude", "ai", "multi-agent", "orchestration", "development"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
    "watchdog>=3.0",
    "pydantic>=2.0",
    "httpx>=0.24",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "mypy>=1.0",
    "ruff>=0.1",
]
dashboard = [
    "fastapi>=0.100",
    "uvicorn>=0.23",
    "jinja2>=3.1",
]

[project.urls]
Homepage = "https://github.com/yourusername/claude-multi-worker-framework"
Documentation = "https://github.com/yourusername/claude-multi-worker-framework/wiki"
Repository = "https://github.com/yourusername/claude-multi-worker-framework"
Issues = "https://github.com/yourusername/claude-multi-worker-framework/issues"

[project.scripts]
cmw = "cmw.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
cmw = ["templates/**/*", "py.typed"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
